+++ Sources/HeidrunCore/Network/HotlineNetworkClient.swift
@@ -584,7 +584,7 @@ public actor HotlineNetworkClient: HotlineClient {
                 name: reply.string(.fileName, encoding: stringEncoding) ?? name,
                 type: typeFCC,
                 creator: creatorFCC,
-                size: size,
+                size: UInt64(size),
                 itemCount: 0
             ),
             creationDate: reply.date(.fileCreationDate),
